+++ packages/antd/src/components/buttons/create/index.tsx
@@ -1,0 +1,78 @@
+import React from "react";
+import { Button, ButtonProps } from "antd";
+import { PlusSquareOutlined } from "@ant-design/icons";
+import {
+    useNavigation,
+    useRouterContext,
+    useTranslate,
+    useCan,
+    useResourceWithRoute,
+    ResourceRouterParams,
+} from "@pankod/refine-core";
+
+export type CreateButtonProps = ButtonProps & {
+    resourceName?: string;
+    hideText?: boolean;
+    ignoreAccessControlProvider?: boolean;
+};
+
+/**
+ * <CreateButton> uses Ant Design's {@link https://ant.design/components/button/ `<Button> component`}.
+ * It uses the {@link https://refine.dev/docs/api-references/hooks/navigation/useNavigation#create `create`} method from {@link https://refine.dev/docs/api-references/hooks/navigation/useNavigation `useNavigation`} under the hood.
+ * It can be useful to redirect the app to the create page route of resource}.
+ *
+ * @see {@link https://refine.dev/docs/api-references/components/buttons/create-button} for more details.
+ */
+export const CreateButton: React.FC<CreateButtonProps> = ({
+    resourceName: propResourceName,
+    hideText = false,
+    ignoreAccessControlProvider = false,
+    children,
+    ...rest
+}) => {
+    const resourceWithRoute = useResourceWithRoute();
+
+    const translate = useTranslate();
+
+    const { create } = useNavigation();
+
+    const { useParams } = useRouterContext();
+
+    const { resource: routeResourceName } = useParams<ResourceRouterParams>();
+
+    const resource = resourceWithRoute(routeResourceName);
+
+    const resourceName = propResourceName ?? resource.name;
+
+    const onButtonClick = () => create(resourceName, "push");
+
+    const { data } = useCan({
+        resource: resourceName,
+        action: "create",
+        queryOptions: {
+            enabled: !ignoreAccessControlProvider,
+        },
+    });
+
+    const createButtonDisabledTitle = () => {
+        if (data?.can) return "";
+        else if (data?.reason) return data.reason;
+        else
+            return translate(
+                "buttons.notAccessTitle",
+                "You don't have permission to access",
+            );
+    };
+
+    return (
+        <Button
+            onClick={onButtonClick}
+            icon={<PlusSquareOutlined />}
+            disabled={data?.can === false}
+            title={createButtonDisabledTitle()}
+            {...rest}
+        >
+            {!hideText && (children ?? translate("buttons.create", "Create"))}
+        </Button>
+    );
+};
