+++ packages/antd/src/components/buttons/list/index.tsx
@@ -1,30 +1,30 @@
-import React, { FC } from "react";
+import React from "react";
 import { Button, ButtonProps } from "antd";
-import { PlusSquareOutlined } from "@ant-design/icons";
-
+import { BarsOutlined } from "@ant-design/icons";
 import {
+    useCan,
     useNavigation,
+    useResourceWithRoute,
     useRouterContext,
     useTranslate,
-    useCan,
-    useResourceWithRoute,
-} from "@hooks";
-import { ResourceRouterParams } from "../../../interfaces";
+    ResourceRouterParams,
+    userFriendlyResourceName,
+} from "@pankod/refine-core";
 
-export type CreateButtonProps = ButtonProps & {
+type ListButtonProps = ButtonProps & {
     resourceName?: string;
     hideText?: boolean;
     ignoreAccessControlProvider?: boolean;
 };
 
 /**
- * <CreateButton> uses Ant Design's {@link https://ant.design/components/button/ `<Button> component`}.
- * It uses the {@link https://refine.dev/docs/api-references/hooks/navigation/useNavigation#create `create`} method from {@link https://refine.dev/docs/api-references/hooks/navigation/useNavigation `useNavigation`} under the hood.
- * It can be useful to redirect the app to the create page route of resource}.
+ * `<ListButton>` is using Ant Design's {@link https://ant.design/components/button/ `<Button>`} component.
+ * It uses the  {@link https://refine.dev/docs/api-references/hooks/navigation/useNavigation#list `list`} method from {@link https://refine.dev/docs/api-references/hooks/navigation/useNavigation `useNavigation`} under the hood.
+ * It can be useful when redirecting the app to the list page route of resource}.
  *
- * @see {@link https://refine.dev/docs/api-references/components/buttons/create-button} for more details.
+ * @see {@link https://refine.dev/docs/api-references/components/buttons/list-button} for more details.
  */
-export const CreateButton: FC<CreateButtonProps> = ({
+export const ListButton: React.FC<ListButtonProps> = ({
     resourceName: propResourceName,
     hideText = false,
     ignoreAccessControlProvider = false,
@@ -33,9 +33,8 @@ export const CreateButton: FC<CreateButtonProps> = ({
 }) => {
     const resourceWithRoute = useResourceWithRoute();
 
+    const { list } = useNavigation();
     const translate = useTranslate();
-
-    const { create } = useNavigation();
 
     const { useParams } = useRouterContext();
 
@@ -45,11 +44,9 @@ export const CreateButton: FC<CreateButtonProps> = ({
 
     const resourceName = propResourceName ?? resource.name;
 
-    const onButtonClick = () => create(resourceName, "push");
-
     const { data } = useCan({
         resource: resourceName,
-        action: "create",
+        action: "list",
         queryOptions: {
             enabled: !ignoreAccessControlProvider,
         },
@@ -67,13 +64,21 @@ export const CreateButton: FC<CreateButtonProps> = ({
 
     return (
         <Button
-            onClick={onButtonClick}
-            icon={<PlusSquareOutlined />}
+            onClick={(): void => list(resourceName, "push")}
+            icon={<BarsOutlined />}
             disabled={data?.can === false}
             title={createButtonDisabledTitle()}
             {...rest}
         >
-            {!hideText && (children ?? translate("buttons.create", "Create"))}
+            {!hideText &&
+                (children ??
+                    translate(
+                        `${resourceName}.titles.list`,
+                        userFriendlyResourceName(
+                            resource.label ?? resourceName,
+                            "plural",
+                        ),
+                    ))}
         </Button>
     );
 };
